Adapt Java Samples and doc's to commhandler version 1.1.0. Move to Java 11.
