+++ SampleCommunicator/src/main/java/ch/smartgridready/communicator/example/AsynchronousSampleCommunicatorTest.java
@@ -5,10 +5,10 @@ import communicator.async.process.Processor;
 import communicator.async.process.ReadExec;
 import communicator.async.process.Sequence;
 import communicator.async.process.WriteExec;
-import communicator.common.api.Float32Value;
-import communicator.common.api.Float64Value;
-import communicator.common.api.StringValue;
-import communicator.common.api.Value;
+import communicator.common.api.values.Float32Value;
+import communicator.common.api.values.Float64Value;
+import communicator.common.api.values.StringValue;
+import communicator.common.api.values.Value;
 import communicator.common.runtime.GenDriverException;
 import communicator.common.runtime.GenDriverModbusException;
 import communicator.common.runtime.GenDriverSocketException;
