+++ SampleCommunicator/src/main/java/ch/smartgridready/communicator/example/WagoSmartMeterCommunicator.java
@@ -1,6 +1,5 @@
 /**
 *Copyright(c) 2022 Verein SmartGridready Switzerland
-* 
 This Open Source Software is BSD 3 clause licensed:
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
@@ -56,7 +55,7 @@ public class WagoSmartMeterCommunicator {
 
 		try {	
 			
-			DeviceDescriptionLoader<DeviceFrame> loader = new DeviceDescriptionLoader<>();
+			DeviceDescriptionLoader loader = new DeviceDescriptionLoader();
 			DeviceFrame tstMeter = loader.load( XML_BASE_DIR, "SGr_04_0014_0000_WAGO_SmartMeterV0.2.1.xml");
 			
 			GenDriverAPI4ModbusRTU mbRTU = new GenDriverAPI4ModbusRTU();
@@ -162,7 +161,6 @@ public class WagoSmartMeterCommunicator {
 		catch ( Exception e)
 		{
 			LOG.error( "Error reading value from device.", e);
-			e.printStackTrace();
 		}
 	}
 }
