+++ doc/img/SGr-Architecture-Overview.drawio
