+++ doc/img/SGr-Architecture-Overview.png
